// lib/types/types.d.ts
export type BaseNode = ast.Node;

export interface InverseChain extends ast.HasLocation {
  strip: StripFlags;
  program: Program;
  chain?: boolean;
}

export interface Part {
  part: string;
  original: string;
  separator?: string;
}

export interface Program {
  type: 'Program';
  /**
   * The root node of a program has no `loc` if it's empty.
   */
  loc: SourceLocation | undefined;
  blockParams?: string[];
  body: Statement[];
  chained?: boolean;
  strip: StripFlags;
}
// ...
export type Statement =
  | MustacheStatement
  | Content
  | BlockStatement
  | PartialStatement
  | PartialBlockStatement;

export interface MustacheStatement extends BaseNode {
  type: 'Decorator' | 'MustacheStatement';
  path: Expression;
  params: Expression[];
  hash: Hash;
  escaped: boolean;
  strip: StripFlags;
}

export interface PathExpression extends BaseNode {
  readonly original: string;
  readonly this: boolean;
  readonly data: boolean;
  readonly depth: number;
  readonly parts: (string | SubExpression)[];
  readonly head: string | SubExpression | undefined;
  readonly tail: string[];
}

export interface SubExpression extends BaseNode {
  readonly original: string;
}

export interface Hash {
  readonly pairs: HashPair[];
}

export interface StripFlags {
  readonly open?: boolean;
  readonly close?: boolean;
  readonly openStandalone?: boolean;
  readonly closeStandalone?: boolean;
  readonly inlineStandalone?: boolean;
}

export interface HashPair {
  readonly key: string;
  readonly value: Expression;
}

export interface ParserPart {
  readonly part: string;
  readonly original: string;
  readonly separator: string;
}

export interface Content extends BaseNode {
  type: 'ContentStatement';
  original: string;
  value: string;
}

export type Expression = SubExpression | PathExpression;

export interface SourcePosition {
  line: number;
  column: number;
}

export interface SourceLocation {
  source: string | undefined;
  start: SourcePosition;
  end: SourcePosition;
}

export interface CallNode {
  path: ast.Expr;
  params: ast.Expr[];
  hash: ast.Hash;
}

export interface OpenPartial {
  strip: ast.StripFlags;
}

export interface OpenPartialBlock extends CallNode {
  strip: ast.StripFlags;
}

export interface OpenRawBlock extends CallNode, BaseNode {}

export interface OpenBlock extends CallNode {
  open: string;
  blockParams: string[];
  strip: StripFlags;
}

export interface OpenInverse extends CallNode {
  blockParams: string[];
  strip: StripFlags;
}

export interface CloseBlock {
  readonly path: PathExpression;
  strip: StripFlags;
}
// ...
export interface LocInfo {
  first_line: number;
  first_column: number;
  last_line: number;
  last_column: number;
}